refactor: Swap channels by their absolute position using index offsets
This fixes the behavior where if a channel was at position 50, and there are no channels with position 49, the channel would not be moved. This uses the index of the channels fetched from rest in order to get their index positions, and then using their absolute positions to swap it.